Preliminary complete lecture/lab schedule
Trying to get a sense of how to balance time across the full range of topics, especially given the number of sessions and how they fall across the calendar. Obviously this is all up for discussion. Scroll down to the bottom for some thoughts about domain-specific modules and lunch speakers.